+++ ui/src/app/dashboard/components/panel/dashboard-panel.component.ts
@@ -103,7 +103,6 @@ import { MatCheckbox } from '@angular/material/checkbox';
         MatFormField,
         MatInput,
         FormsModule,
-        MatError,
         MatRadioGroup,
         MatRadioButton,
         MatCheckbox,
